chore(update-api): backport of 4144, Inbound instances API rework
