+++ bundle/default-bundle/src/test/java/io/camunda/connector/runtime/app/CorsConfiguration.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.camunda.connector.runtime.app;
+
+import java.util.List;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@Profile("local")
+public class CorsConfiguration {
+  @Value("${camunda.endpoints.cors.allowed-origins:}")
+  private String allowedOrigins;
+
+  @Value("${camunda.endpoints.cors.mappings:}")
+  private String mappings;
+
+  @Bean
+  public WebMvcConfigurer corsConfigurer() {
+    return new WebMvcConfigurer() {
+      @Override
+      public void addCorsMappings(CorsRegistry registry) {
+        if (StringUtils.isNotBlank(allowedOrigins) && StringUtils.isNotBlank(mappings)) {
+          String[] allowedOriginsArray = allowedOrigins.split(",");
+          List<String> mappingsList = List.of(mappings.split(","));
+          mappingsList.forEach(
+              mapping ->
+                  registry
+                      .addMapping(mapping)
+                      .allowedOrigins(allowedOriginsArray)
+                      .allowedMethods("*"));
+        }
+      }
+    };
+  }
+}
